requirements.txt
# Requirements

# Dependências principais
langchain>=0.1.0
langchain-community>=0.0.10
llama-index>=0.9.0
pydantic>=2.0.0
PyYAML>=6.0
requests>=2.28.0
python-dotenv>=1.0.0

# Dependências para CI/CD
gitpython>=3.1.30
semantic-version>=2.10.0
jinja2>=3.1.2

# Dependências para IaC
terraform-parser>=0.1.0
cfn-lint>=0.77.0
ansible-lint>=6.14.0
kubernetes>=26.1.0

# Dependências para modelos de IA
ollama>=0.1.0
transformers>=4.30.0
torch>=2.0.0
sentence-transformers>=2.2.2

# Dependências para utilidades
rich>=13.3.0
typer>=0.9.0
